Condense form logic
