+++ src/views/GovernanceReviewTeam/Actions/RequestEdits.tsx
@@ -1,22 +1,28 @@
 import React from 'react';
+import { FormProvider, useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 
-import useActionForm from './ActionFormWrapper/useActionForm';
+import ActionForm, { SystemIntakeActionFields } from './ActionFormWrapper';
 
 const RequestEdits = ({ systemIntakeId }: { systemIntakeId: string }) => {
   const { t } = useTranslation('action');
 
-  const { ActionForm, handleSubmit } = useActionForm({ systemIntakeId });
+  const form = useForm<SystemIntakeActionFields>();
+
+  const { handleSubmit } = form;
 
   const submit = handleSubmit(() => null);
 
   return (
-    <ActionForm
-      title={t('requestEdits.title')}
-      description={t('requestEdits.description')}
-      breadcrumbItems={[{ text: t('requestEdits.breadcrumb') }]}
-      onSubmit={submit}
-    />
+    <FormProvider<SystemIntakeActionFields> {...form}>
+      <ActionForm
+        systemIntakeId={systemIntakeId}
+        title={t('requestEdits.title')}
+        description={t('requestEdits.description')}
+        breadcrumb={t('requestEdits.breadcrumb')}
+        onSubmit={submit}
+      />
+    </FormProvider>
   );
 };
